src/main/java/org/github/mbwardell/classhunter/Jar.java
/*
Copyright 2021 M. Wardell
        Use of this source code is governed by an MIT-style
        license that can be found in the LICENSE file or at
        https://opensource.org/licenses/MIT
*/
package org.github.mbwardell.classhunter;

import java.io.IOException;
import java.nio.file.Files;
import java.nio.file.Path;
import java.util.ArrayList;
import java.util.List;
import java.util.zip.ZipEntry;
import java.util.zip.ZipInputStream;

public class Jar {
    private final Path pathToJar;

    public Jar(Path pathToJar) {
        this.pathToJar = pathToJar;
    }

    public List<String> getClasses() throws IOException {
        List<String> classNames = new ArrayList<>();
        try (ZipInputStream zip = new ZipInputStream(Files.newInputStream(pathToJar))) {
            for (ZipEntry entry = zip.getNextEntry(); entry != null; entry = zip.getNextEntry()) {
                if (!entry.isDirectory() && entry.getName().endsWith(".class")) {
                    String className = entry.getName().replace('/', '.'); // including ".class"
                    classNames.add(className.substring(0, className.length() - ".class".length()));
                }
            }
        }
        return classNames;
    }
}
